backend/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
django-environ = "^0.12.0"
django-simple-captcha = "^0.6.1"
